Move GRIDSS from gatk_only somatic pipelines
