+++ janis_pipelines/wgs_somatic/wgssomatic_variantsonly.py
@@ -5,6 +5,7 @@ from janis_bioinformatics.tools.bcftools import BcfToolsSort_1_9
 from janis_bioinformatics.tools.common import GATKBaseRecalBQSRWorkflow_4_1_3
 from janis_bioinformatics.tools.gatk4 import Gatk4GatherVcfs_4_1_3
 from janis_bioinformatics.tools.htslib import BGZipLatest
+from janis_bioinformatics.tools.papenfuss import Gridss_2_6_2
 from janis_bioinformatics.tools.pmac import (
     CombineVariants_0_0_8,
     GenerateVardictHeaderLines,
@@ -80,6 +81,32 @@ class WGSSomaticMultiCallersVariantsOnly(WGSSomaticGATKVariantsOnly):
                 "The threshold for VarDict's allele frequency, default: 0.05 or 5%",
                 quality=InputQualityType.configuration,
             ),
+        )
+
+    def add_gridss(self, normal_bam_source, tumor_bam_source):
+
+        # GRIDSS
+        self.step(
+            "vc_gridss",
+            Gridss_2_6_2(
+                bams=[normal_bam_source, tumor_bam_source],
+                reference=self.reference,
+                blacklist=self.gridss_blacklist,
+            ),
+        )
+
+        # GRIDSS
+        self.output(
+            "out_gridss_assembly",
+            source=self.vc_gridss.assembly,
+            output_folder="gridss",
+            doc="Assembly returned by GRIDSS",
+        )
+        self.output(
+            "out_variants_gridss",
+            source=self.vc_gridss.out,
+            output_folder="gridss",
+            doc="Variants from the GRIDSS variant caller",
         )
 
     def add_gatk_variantcaller(self, normal_bam_source, tumor_bam_source):
