+++ janis_pipelines/wgs_somatic_gatk/wgssomaticgatk.py
@@ -53,9 +53,7 @@ class WGSSomaticGATK(WGSSomaticGATKVariantsOnly):
     def constructor(self):
         self.add_inputs()
         self.add_preprocessing_steps()
-        self.add_gridss(
-            normal_bam_source=self.normal.out_bam, tumor_bam_source=self.tumor.out_bam
-        )
+
         self.add_gatk_variantcaller(
             normal_bam_source=self.normal.out_bam, tumor_bam_source=self.tumor.out_bam
         )
